requirements.txt
anyio==3.6.2
asyncpg==0.27.0
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==41.0.3
ecdsa==0.18.0
Faker==18.5.1
fastapi==0.95.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
passlib==1.7.4
psycopg2-binary==2.9.6
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.7
python-dateutil==2.8.2
python-jose==3.3.0
python-multipart==0.0.6
requests==2.28.2
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.10
starlette==0.26.1
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
